Refactor choice builders for reuse in tests
